📚 Improve README significantly
Having all the documentation on a separate website is a bit overkill.
It is more convenient for users to have it all in the README.

This PR recreates the website's content in the README, with some minor
changes.

The website is still available, but the plan is to shut it down at some
point.