🐛 Fix data race in end-to-end tests (#66)
Most end-to-end tests are stateless, but some are not. For instance,
end-to-end tests that use a Terraform Cloud backend share that backend.
Running multiple instances of these tests concurrently results in data
races and flakiness.

This PR changes our end-to-end tests to ensure we don't change the state
used by these tests. The original state is set up manually beforehand.